add new micro command
