+++ fil-proofs-tooling/Cargo.toml
@@ -21,3 +21,5 @@ glob = "0.3"
 human-size = "0.4"
 prettytable-rs = "0.8"
 regex = "1.1.6"
+prometheus = "0.6.1"
+commandspec = "0.12.2"
